superfrog-scheduler/src/main/java/edu/tcu/cs/superfrogscheduler/appointment/AppointmentService.java
package edu.tcu.cs.superfrogscheduler.appointment;

import jakarta.transaction.Transactional;
import org.springframework.stereotype.Service;

@Service
@Transactional
public class AppointmentService {

    private final AppointmentRepository appointmentRepository;

    public AppointmentService(AppointmentRepository appointmentRepository) {
        this.appointmentRepository = appointmentRepository;
    }

    public Appointment getAppointmentById(String E_id){
        return this.appointmentRepository.findById(E_id).get();
    }

}
